fix regex/validator, add selenium tests
