+++ src/org/labkey/test/util/AuditLogHelper.java
@@ -1,6 +1,7 @@
 package org.labkey.test.util;
 
 import org.jetbrains.annotations.Nullable;
+import org.json.JSONException;
 import org.labkey.remoteapi.CommandException;
 import org.labkey.remoteapi.Connection;
 import org.labkey.remoteapi.query.ContainerFilter;
@@ -10,18 +11,23 @@ import org.labkey.remoteapi.query.SelectRowsResponse;
 import org.labkey.remoteapi.query.Sort;
 import org.labkey.test.Locator;
 import org.labkey.test.WebDriverWrapper;
+import org.labkey.test.WebTestHelper;
 import org.labkey.test.pages.core.admin.ShowAdminPage;
 import org.labkey.test.pages.core.admin.ShowAuditLogPage;
 
 import java.io.IOException;
 import java.net.URL;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Stream;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.fail;
 
 public class AuditLogHelper
 {
@@ -187,5 +193,206 @@ public class AuditLogHelper
     public interface ConnectionSupplier
     {
         Connection get() throws IOException, CommandException;
+    }
+
+    public boolean validateExpectedRowInDomainPropertyAuditLog(List<Map<String, Object>> domainPropertyEventRows, String propertyName, Map<String, String> expectedColumns, @Nullable Map<String, String> expectedComment)
+    {
+        boolean pass = true;
+
+        for(Map<String, Object> row : domainPropertyEventRows)
+        {
+
+            if(getLogColumnValue(row, "propertyname").equals(propertyName))
+            {
+                TestLogger.log("Validate the columns for property '" + propertyName + "'.");
+                for(String fieldName : expectedColumns.keySet())
+                {
+                    if(!getLogColumnValue(row, fieldName).equals(expectedColumns.get(fieldName)))
+                    {
+                        pass = false;
+                        TestLogger.log("************** For field '" + fieldName + "' expected value '" + expectedColumns.get(fieldName) + "' found '" + row.get(fieldName) + "' **************");
+                    }
+                }
+
+                if(null != expectedComment)
+                {
+                    TestLogger.log("Validate that the Comment field is as expected.");
+                    Map<String, String> commentFieldValues = getDomainPropertyEventComment(row);
+                    pass = validateCommentHasExpectedValues(commentFieldValues, expectedComment) && pass;
+                }
+            }
+
+        }
+
+        return pass;
+    }
+
+    private boolean validateCommentHasExpectedValues(Map<String, String> comment, Map<String, String> expected)
+    {
+        boolean pass = true;
+
+        for(String key : expected.keySet())
+        {
+            if(!expected.get(key).equals(comment.get(key)))
+            {
+                TestLogger.log("************** Comment value does not contain expected value for field '" + key + "'. Expected '" + expected.get(key) + "' found '" + comment.get(key) + "'.  **************");
+                pass = false;
+            }
+        }
+
+        return pass;
+    }
+
+    public List<Map<String, Object>> getDomainPropertyEventsFromDomainEvents(String projectName, String domainName, @Nullable List<String> ignoreIds)
+    {
+        List<String> domainEventIds = getDomainEventIds(projectName, domainName);
+
+        if(null != ignoreIds)
+        {
+            TestLogger.log("Removing the ignore ids from the list.");
+            domainEventIds.removeAll(ignoreIds);
+        }
+
+        TestLogger.log("Get all of the Domain Property Events for '" + domainName + "' that are linked to the domain events.");
+        List<Map<String, Object>> domainPropertyEventRows = getDomainPropertyEventLog(domainName, domainEventIds);
+        TestLogger.log("Number of 'Domain Property Event' log entries: " + domainPropertyEventRows.size());
+
+        return domainPropertyEventRows;
+    }
+
+    private List<String> getDomainEventIds(String projectName, String domainName)
+    {
+        TestLogger.log("Get a list of the Domain Events for project '" + projectName + "'. ");
+        List<Map<String, Object>> domainAuditEventAllRows = getDomainEventLog(projectName);
+        TestLogger.log("Number of 'Domain Event' log entries for '" + projectName + "': " + domainAuditEventAllRows.size());
+
+        TestLogger.log("Filter the list to look only at '" + domainName + "'.");
+        List<Map<String, Object>> domainAuditEventRows = new ArrayList<>();
+
+        for(Map<String, Object> row : domainAuditEventAllRows)
+        {
+            if(getLogColumnValue(row, "domainname").toLowerCase().trim().equals(domainName.toLowerCase().trim()))
+                domainAuditEventRows.add(row);
+        }
+
+        List<String> domainEventIds = new ArrayList<>();
+        domainAuditEventRows.forEach((event)->domainEventIds.add(getLogColumnValue(event, "rowid")));
+
+        TestLogger.log("Number of 'Domain Event' log entries for '" + domainName + "': " + domainEventIds.size());
+
+        return domainEventIds;
+    }
+
+    public List<String> getDomainEventIdsFromPropertyEvents(List<Map<String, Object>> domainPropertyEventRows)
+    {
+        List<String> domainEventIds = new ArrayList<>();
+
+        for(Map<String, Object> row : domainPropertyEventRows)
+        {
+            domainEventIds.add(getLogColumnValue(row, "domaineventid"));
+        }
+
+        return domainEventIds;
+    }
+
+    private List<Map<String, Object>> getDomainEventLog(String projectName)
+    {
+        Connection cn = WebTestHelper.getRemoteApiConnection();
+        SelectRowsCommand cmd = new SelectRowsCommand("auditLog", "DomainAuditEvent");
+        cmd.setRequiredVersion(9.1);
+        cmd.setColumns(Arrays.asList("rowid", "created", "createdby", "impersonatedby", "projectid", "domainuri", "domainname", "comment"));
+        cmd.addFilter("projectid/DisplayName", projectName, Filter.Operator.EQUAL);
+        cmd.setContainerFilter(ContainerFilter.AllFolders);
+
+        return executeSelectCommand(cn, cmd);
+    }
+
+    private List<Map<String, Object>> getDomainPropertyEventLog(String domainName, @Nullable List<String> eventIds)
+    {
+        Connection cn = WebTestHelper.getRemoteApiConnection();
+        SelectRowsCommand cmd = new SelectRowsCommand("auditLog", "DomainPropertyAuditEvent");
+        cmd.setRequiredVersion(9.1);
+        cmd.setColumns(Arrays.asList("Created", "CreatedBy", "ImpersonatedBy", "propertyname", "action", "domainname", "domaineventid", "Comment"));
+        cmd.addFilter("domainname", domainName, Filter.Operator.EQUAL);
+
+        if(null != eventIds)
+        {
+            StringBuilder stringBuilder = new StringBuilder();
+            eventIds.forEach((id)->{
+                if(!stringBuilder.isEmpty())
+                    stringBuilder.append(";");
+                stringBuilder.append(id);
+            });
+            cmd.addFilter("domaineventid/rowid", stringBuilder, Filter.Operator.IN);
+        }
+
+        cmd.setContainerFilter(ContainerFilter.AllFolders);
+
+        return executeSelectCommand(cn, cmd);
+    }
+
+    private List<Map<String, Object>> executeSelectCommand(Connection cn, SelectRowsCommand cmd)
+    {
+        List<Map<String, Object>> rowsReturned = new ArrayList<>();
+        try
+        {
+            SelectRowsResponse response = cmd.execute(cn, "/");
+            TestLogger.log("Number of rows: " + response.getRowCount());
+            rowsReturned.addAll(response.getRows());
+        }
+        catch(IOException | CommandException ex)
+        {
+            // Just fail here, don't toss the exception up the stack.
+            fail("There was a command exception when getting the log: " + ex);
+        }
+
+        return rowsReturned;
+    }
+
+    private Map<String, String> getDomainPropertyEventComment(Map<String, Object> row)
+    {
+        String comment = getLogColumnValue(row, "Comment");
+
+        String[] commentAsArray = comment.split(";");
+
+        Map<String, String> fieldComments = new HashMap<>();
+
+        for (String s : commentAsArray)
+        {
+            String[] fieldValue = s.split(":");
+
+            // If the split on the ':' produced more than two entries in the array it most likely means that the
+            // comment for that property had a : in it. So treat the first entry as the field name and then concat the
+            // other fields together.
+            // For example the ConditionalFormats field will log the following during an update:
+            // ConditionalFormats: old: <none>, new: 1;
+            // And a create of a Lookup will log as:
+            // Lookup: [Schema: lists, Query: LookUp01];
+            StringBuilder sb = new StringBuilder();
+            sb.append(fieldValue[1].trim());
+
+            for (int j = 2; j < fieldValue.length; j++)
+            {
+                sb.append(":");
+                sb.append(fieldValue[j]);
+            }
+
+            fieldComments.put(fieldValue[0].trim(), sb.toString());
+        }
+
+        return fieldComments;
+    }
+
+    private String getLogColumnValue(Map<String, Object> rowEntry, String columnName)
+    {
+        try
+        {
+            return ((Map<String, Object>) rowEntry.get(columnName)).get("value").toString();
+        }
+        catch(JSONException je)
+        {
+            // Just fail here, don't toss the exception up the stack.
+            throw new IllegalArgumentException(je);
+        }
     }
 }
